requirements.txt
cirq-core>=0.15.0
cirq-google>=0.15.0
# When changing Cirq requirements be sure to update dev_tools/write-ci-requirements.py

scipy
ipython
black
pylint
pytest
